requirements.txt
python-dotenv
requests
geopy
pyowm
nltk
wikipedia-api
beautifulsoup4
spotipy
gtts
playsound
pyaudio
speechrecognition
keyboard
torch
numpy
openai
deep_translator
pyttsx3